+++ src/file_pipe_log/format.rs
@@ -178,11 +178,15 @@ impl LogFileContext {
         }
     }
 
+    /// Return the `signature` in `u32` format.
+    ///
+    /// Here, in mainly common case in producion, the `Config::purge_threshold` is
+    /// always limited to less than `UINT32_MAX` * `Config::target_file_size`,
+    /// the higher 32 bit keeps always same with zero bits. That is, the count of
+    /// files will be limited to less than `UINT32_MAX`. So, we just use the low
+    /// 32 bit as the `signature` by default.
     pub fn get_signature(&self) -> u32 {
-        let high = (self.id.seq as u64 >> 32) as u32;
-        let low = self.id.seq as u32;
-        let flag = self.id.queue as u32;
-        high ^ low ^ flag
+        self.id.seq as u32
     }
 }
 
@@ -232,5 +236,17 @@ mod tests {
         assert_eq!(header2.version().to_u64(), header1.version().to_u64());
         let header3 = LogFileFormat::from_version(Version::default());
         assert_eq!(header3.version(), header1.version());
+    }
+
+    #[test]
+    fn test_file_context() {
+        let mut file_context = LogFileContext::dummy(LogQueue::Append);
+        assert_eq!(file_context.get_signature(), 0);
+        file_context.id.seq = 10;
+        assert_eq!(file_context.get_signature(), 10);
+        let abnormal_seq = (file_context.id.seq << 32) as u64 + 100 as u64;
+        file_context.id.seq = abnormal_seq;
+        assert_ne!(file_context.get_signature() as u64, abnormal_seq);
+        assert_eq!(file_context.get_signature(), 100);
     }
 }