+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+* Support recycling obsolete logs in Raft Engine to reduce the cost on `fallocate` with new files #224 
+
 ### Bug Fixes
 
 * Unconditionally tolerate `fallocate` failures as a fix to its portability issue. Errors other than `EOPNOTSUPP` will still emit a warning.
@@ -24,7 +26,6 @@
 
 ### New Features
 
-* Support recycling obsolete logs in Raft Engine to reduce the cost on `fallocate` with new files #224 
 * Support limiting the memory usage of Raft Engine under new feature `swap` #211 
 * Add a new Prometheus counter `raft_engine_memory_usage` to track memory usage #207